Schema changes, start to format java args using JCOMMANDER_OPTS, and some tighter type safety.
A big change was in selectInputsFieldsAsExpressionRecord.  That function now takes a list of the keys in the field that values are being extracted from.  That list MUST match the list of keys present in the object.  I needed to pass it as an argument to capture the values for the runtime so that I could _only_ include the keys from the object rather than try to select fields that wouldn't be there.  In doing that refactoring, there was some more work to get more attention on default values.  If a value doesn't have a default on the template, the field must be required on the object that we're scraping the fields from.  That required me to go through and set a number of defaults within the schema.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>